# public/swagger.yaml
openapi: 3.0.0
info:
  description: API REST basado en Greenter - Facturación Electrónica.
  version: "1.3"
  title: Lycet API
  contact:
    name: Giancarlos Salas
    email: giansalex@gmail.com
...
security:
  - ApiToken: []
paths:
...
  /despatch/send:
    post:
      tags:
        - despatch
      summary: Enviar una Guía de Remisión
      description: Retorna la respuesta de SUNAT y el XML firmado
      operationId: sendDespatch
      requestBody:
        $ref: "#/components/requestBodies/Despatch"
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResponse"
        "400":
          description: Error de Validación
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /despatch/xml:
    post:
      tags:
        - despatch
      summary: Generar el XML de una Guía de Remisión
      description: Retorna el XML
      operationId: getDespatchXml
      requestBody:
        $ref: "#/components/requestBodies/Despatch"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            text/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /despatch/status:
    get:
      tags:
        - despatch
      summary: Obtiene el estado del ticket
      description: Retorna el CDR
      operationId: getDespatchStatus
      parameters:
        - name: ticket
          in: query
          description: Ticket del comprobante
          schema:
            type: string
        - name: ruc
          in: query
          description: RUC de empresa (Multiples Empresas)
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResult"
  /despatch/pdf:
    post:
      tags:
        - despatch
      summary: Generar el Pdf de una Guía de Remisión
      description: Retorna el Pdf
      operationId: getDespatchPdf
      requestBody:
        $ref: "#/components/requestBodies/Despatch"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            application/Pdf:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
...
servers:
  - url: http://lycet.api/api/v1
  - url: https://lycet.api/api/v1
components:
  requestBodies:
...
    Despatch:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Despatch"
      description: Comprobante
      required: true
...
  securitySchemes:
    ApiToken:
      name: token
      type: apiKey
      in: query
      description: Api Token
  schemas:
    Address:
      type: object
      properties:
        ubigueo:
          type: string
        codigoPais:
          type: string
        departamento:
          type: string
        provincia:
          type: string
        distrito:
          type: string
        urbanizacion:
          type: string
        direccion:
          type: string
        codLocal:
          type: string
    Client:
      type: object
      properties:
        tipoDoc:
          type: string
        numDoc:
          type: string
        rznSocial:
          type: string
        address:
          $ref: "#/components/schemas/Address"
        email:
          type: string
        telephone:
          type: string
    Company:
      type: object
      properties:
        ruc:
          type: string
        razonSocial:
          type: string
        nombreComercial:
          type: string
        address:
          $ref: "#/components/schemas/Address"
        email:
          type: string
        telephone:
          type: string
...
    DetailAttribute:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        value:
          type: string
        fecInicio:
          type: string
          format: date-time
        fecFin:
          type: string
          format: date-time
        duracion:
          type: integer
...
    Direction:
      type: object
      properties:
        ubigueo:
          type: string
        direccion:
          type: string
...
    Puerto:
      type: object
      properties:
        codigo:
          type: string
        nombre:
          type: string
    Transportist:
      type: object
      properties:
        tipoDoc:
          type: string
        numDoc:
          type: string
        rznSocial:
          type: string
        nroMtc:
          type: string
        placa:
          type: string
        choferTipoDoc:
          type: string
        choferDoc:
          type: string
    Vehicle:
      type: object
      properties:
        placa:
          type: string
        nroCirculacion:
          type: string
        codEmisor:
          type: string
        nroAutorizacion:
          type: string
        secundarios:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
    Driver:
      type: object
      properties:
        tipo:
          type: string
        tipoDoc:
          type: string
        nroDoc:
          type: string
        nombres:
          type: string
        apellidos:
          type: string
        licencia:
          type: string
    Shipment:
      type: object
      properties:
        codTraslado:
          type: string
        desTraslado:
          type: string
        sustentoPeso:
          type: string
        indTransbordo:
          type: boolean
        indicadores:
          type: array
          items:
            type: string
        pesoItems:
          type: number
          format: float
        pesoTotal:
          type: number
          format: float
        undPesoTotal:
          type: string
        numBultos:
          type: integer
        modTraslado:
          type: string
        fecTraslado:
          type: string
          format: date-time
        numContenedor:
          type: string
        contenedores:
          type: array
          items:
            type: string
        codPuerto:
          type: string
        puerto:
          $ref: '#/components/schemas/Puerto'
        aeropuerto:
          $ref: '#/components/schemas/Puerto'
        transportista:
          $ref: '#/components/schemas/Transportist'
        vehiculo:
          $ref: '#/components/schemas/Vehicle'
        choferes:
          type: array
          items:
            $ref: '#/components/schemas/Driver'
        llegada:
          $ref: '#/components/schemas/Direction'
        partida:
          $ref: '#/components/schemas/Direction'
    AdditionalDoc:
      type: object
      properties:
        tipoDesc:
          type: string
        tipo:
          type: string
        nro:
          type: string
        emisor:
          type: string
    DespatchDetail:
      type: object
      properties:
        codigo:
          type: string
        descripcion:
          type: string
        unidad:
          type: string
        cantidad:
          type: number
          format: float
        codProdSunat:
          type: string
        atributos:
          type: array
          items:
            $ref: '#/components/schemas/DetailAttribute'
    Despatch:
      type: object
      properties:
        version:
          type: string
        tipoDoc:
          type: string
        serie:
          type: string
        correlativo:
          type: string
        observacion:
          type: string
        fechaEmision:
          type: string
          format: date-time
        company:
          $ref: '#/components/schemas/Company'
        destinatario:
          $ref: '#/components/schemas/Client'
        tercero:
          $ref: '#/components/schemas/Client'
        comprador:
          $ref: '#/components/schemas/Client'
        envio:
          $ref: '#/components/schemas/Shipment'
        addDocs:
          type: array
          items:
            $ref: '#/components/schemas/AdditionalDoc'
        details:
          type: array
          items:
            $ref: '#/components/schemas/DespatchDetail'
        name:
          type: string
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    CdrResponse:
      type: object
      properties:
        accepted:
          type: boolean
          readOnly: true
        id:
          type: string
        code:
          type: string
        description:
          type: string
        notes:
          type: array
          items:
            type: string
...
    SummaryResult:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/Error"
        ticket:
          type: string
    StatusResult:
      type: object
      properties:
        cdrZip:
          type: string
          format: byte
        cdrResponse:
          $ref: "#/components/schemas/CdrResponse"
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/Error"
        code:
          type: string
...
    SummaryResponse:
      type: object
      properties:
        xml:
          type: string
        hash:
          type: string
          description: Resumen de la Firma Digital
        sunatResponse:
          $ref: "#/components/schemas/SummaryResult"
    ValidationResponse:
      type: object
      properties:
        message:
          type: string
        field:
          type: string
...